fs/namespace: add flag to disable propagation to sys_move_mount
If MOVE_MOUNT_NO_PROPAGATE flag provided to sys_move_mount, then newly
created mount will not propagate even if it's new parent is shared. Also
it won't change it's shared and master ids (e.g.: what was private will
remain private).

One of the reasons why this is needed is to be able to recreate
different complex mount trees in CRIU so that propagation remains the
same after a checkpoint/restore.

I've talked on LPC about different problems we face with mounts in CRIU:

https://linuxplumbersconf.org/event/7/contributions/640/attachments/547/968/CRIU_mounts_migration__problems_and_solutions.pdf

And the propagation is the most important one.

Before the patch there was a way to do effectively the same thing by
creating mount with propagation, make it private (or in contrary make
all other mounts in propagation group private) and lock it with children
and later umount all other mounts of propagation. But inheriting proper
shared ids is a big pain here because propagation can change your ids.
Also there are some cases when propagations overmount original mount and
you can't lock it easily. So we need a simple way to recreate it.

Here is a small utility to illustrate how it works:
https://github.com/Snorch/move-mount-no-propagate/

Signed-off-by: Pavel Tikhomirov <ptikhomirov@virtuozzo.com>